decimal issue fix

+++ src/main/java/com/asg/finance/dto/BankPayFdaItemDto.java
@@ -1,11 +1,12 @@
 package com.asg.finance.dto;
 
+import java.math.BigDecimal;
 import lombok.Data;
 
 @Data
 public class BankPayFdaItemDto {
     private Long chargePoid;
-    private Double pdaAmount;
+    private BigDecimal pdaAmount;
     private String remarks;
     private String refDocId;
     private String refDocPoid;
