+++ src/main/java/com/asg/finance/dto/BankPaymentChargeDetailResponse.java
@@ -2,6 +2,7 @@ package com.asg.finance.dto;
 
 import lombok.Data;
 import com.asg.common.lib.dto.LovGetListDto;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -11,7 +12,7 @@ public class BankPaymentChargeDetailResponse {
     private Long detRowId;
     private Long chargePoid;
     private LovGetListDto chargeDet;
-    private Long chargeAmount;
+    private BigDecimal chargeAmount;
     private String description;
     private String remarks;
     private String refDocId;
@@ -19,8 +20,8 @@ public class BankPaymentChargeDetailResponse {
     private LovGetListDto refDocDet;
     private Long fdaDetRowId;
     private String checkAll;
-    private Long pdaAmount;
-    private Long ffAmount;
+    private BigDecimal pdaAmount;
+    private BigDecimal ffAmount;
     private String createdBy;
     private LocalDateTime createdDate;
 }
